Combine BlockArray tests
